Refactor jwt Utils
