+++ ccgame-service/src/main/java/com/doubleat/ccgame/dto/common/UserDto.java
@@ -1,6 +1,5 @@
 package com.doubleat.ccgame.dto.common;
 
-import com.sun.istack.Nullable;
 import lombok.*;
 
 @Data
