+++ ccgame-service/src/main/java/com/doubleat/ccgame/jwt/JwtAuthenticationFilter.java
@@ -1,5 +1,6 @@
 package com.doubleat.ccgame.jwt;
 
+import com.doubleat.ccgame.utils.JwtService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,14 +22,18 @@ import java.io.IOException;
 
 @Component
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
-    private final UserDetailsService userDetailsService;
-
-    @Autowired
-    public JwtAuthenticationFilter(@Qualifier("userDetailsServiceImpl") UserDetailsService userDetailsService) {
-        this.userDetailsService = userDetailsService;
-    }
 
     private static final Logger logger = LoggerFactory.getLogger(JwtAuthenticationFilter.class);
+
+    private final UserDetailsService userDetailsService;
+
+    private final JwtService jwtService;
+
+    @Autowired
+    public JwtAuthenticationFilter(@Qualifier("userDetailsServiceImpl") UserDetailsService userDetailsService, JwtService jwtService) {
+        this.userDetailsService = userDetailsService;
+        this.jwtService = jwtService;
+    }
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
@@ -36,9 +41,9 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
         if (StringUtils.hasText(authHeader) && authHeader.startsWith("Bearer ")) {
             String jwtToken = authHeader.substring(7);
             try {
-                JwtUtils.validateJwtToken(jwtToken);
+                jwtService.validateJwtToken(jwtToken);
 
-                String username = JwtUtils.getUsernameFromJwtToken(jwtToken);
+                String username = jwtService.getUsernameFromJwtToken(jwtToken);
 
                 UserDetails userDetails = userDetailsService.loadUserByUsername(username);
                 UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
@@ -57,4 +62,5 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
         filterChain.doFilter(request, response);
     }
+
 }
